Update patch set 3
Patch Set 3:

why tests/pthread_test.cpp is not here in patch set 4?

For the tests in tests/euid_test.cpp in patch set 3, we can not test the seteuid in cts, but I think we can still test the euid method, could you add that to unistd?

Patch-set: 3
Label: Verified=0