Update patch set 2
Patch Set 2:

Duplicates won't make it this far because LinkProperties suppresses duplicate DNS servers.

As for ordering, the position that the code takes is that [192.168.0.1, 8.8.8.8, 8.8.4.4] is identical to [8.8.4.4, 8.8.8.8, 192.168.0.1]. Though you could argue that order should matter, I don't think it's an unreasonable position to take.

Added a TODO for both items.

Patch-set: 2
Label: Verified=0