Update patch set 2
Patch Set 2:

My thought was that we should probably be defining it in sys/ucontext.h.  It is defined there for other architectures but missing for aarch64. Glibc defines it in their sys/ucontext.h for aarch64 (although they do define it as an int as you mention, which does seem strange).

Patch-set: 2