Update patch set 1
Patch Set 1:

> I assume the switch is becuase the openbsd version seems to be a
 > bit more lenient about the order of the options. Is that correct?

no, though i was pleased to see that. it's really just that most of our BSD source is OpenBSD and in particular all but three files of our stdio is OpenBSD. this was one of the three, and the only reason we had FreeBSD's was because they've had O_CLOEXEC support for years. OpenBSD only just got it. (all those other O_CLOEXEC-related fixes this afternoon were brand new too.)

Patch-set: 1