Update patch set 1
Patch Set 1:

(2 comments)

> Andrew used to add code static analysis (WITH_SYNTAX_CHECK) to the
 > build system.
 > Not sure if this can be covered by that?

I'm not sure this is useful to anyone outside bionic. What I'm trying to verify is that certain warnings (or errors) get emitted by the compiler when calling certain functions in a specific way.

Aside from not being generally useful for other projects (probably?), I don't think using a similar approach to the static analysis implementation in build/ would work without rather heavy modification, since this depends on being passed all the normal compilation flags, but must be followed by piping to FileCheck, with some args, and the full path to the source file. It might be doable, but I'm not convinced it would be better. I'll keep thinking about it to see if I can convince myself otherwise though.

Patch-set: 1