Update patch set 1
Patch Set 1:

This turned out to be a bug in the kernel's futex implementation. See:

  https://lkml.org/lkml/2014/10/17/336

I have tested the patch in the link above and confirm it works. Below are the details of the tests I carried out.

I ran 5 times the test program which we use to trigger the issue. The first run was ~14 hours long (I ran the test overnight), the other four runs where all ~1 hour long. With a patched kernel none of these runs triggered the bug. I then did the same for an unpatched kernel. The first run gave a hit after about 30 minutes. The second didn't give a hit within 1 hour of execution time. The third gave a hit after ~5 minutes. The fourth gave a hit after ~10 minutes, the fifth after ~20 minutes. In summary, with an unpatched kernel I could reproduce the issue within 30 minutes from the start of the test for 4 out of the 5 tests I ran. I also did a longer test for about 60 hours with a patched kernel. It worked fine.

Many thanks to Orjian Eide, Hans Boehm and Catalin Marinas for their help with this.

Patch-set: 1