Update patch set 1
Patch Set 1:

We don't have a self contained test that we could upstream easily. Sorry.

We could write one, but then this would really be a futex test (testing the futex syscall, rather than pthread_mutex) and should probably be part of the kernel test infrastructure. I guess it is not bionic's responsibility to test that the Linux kernel does really work properly.

Patch-set: 1