Update patch set 1
Patch Set 1: Code-Review-1

I'm not sure this change is useful. The DNS servers are passed down into bionic by ConnectivityService, which performs an order-insensitive comparison.

Therefore, if the order of the DNS servers for a given network changes, the framework will not update bionic, and this function will never be called. So it doesn't really matter if this function is order-sensitive or not.

Patch-set: 1
Reviewer: Lorenzo Colitti <1000835@85c56323-6fa9-3386-8a01-6480fb634889>
Label: Code-Review=-1