Update patch set 4
Patch Set 4:

Dmitry (Elliot?), can you say what comes next to move this along?

A week or so ago it looked like it had taken on a life of its own (sent for review, picked up by bots), but that seems to have stalled.  What is needed to move this forwards?  Thanks.

Patch-set: 4