Update patch set 1
Patch Set 1:

> bionic-presubmit #313 SUCCESS: http://go/bionicbb/job/bionic-presubmit/313/

This change is still not finished, submitting here for review is just to check if this is in the right direction to design the pthread test with PTHREAD_PROCESS_SHARE attribute, so that we can cover the tests in system/extras and remove them. Please help to give more comments and suggestions.

Patch-set: 1