Update patch set 2
Patch Set 2: Code-Review+2

(1 comment)

We'll want to keep an eye on the buildbot after this is submitted to see if this has any effect on the XML test report generation or coverage. I think we should be fine on the former, but the solution for the latter is probably just to use this sparingly.

Patch-set: 2
Label: Code-Review=+2