Update patch set 10
Patch Set 10: Code-Review+2

i still think a "run this code with a pthread_t running and passed in to it" would be a useful and more general primitive, but this is fine for now. we can always come back for that.

Patch-set: 10
Reviewer: Elliott Hughes <1003224@85c56323-6fa9-3386-8a01-6480fb634889>
Label: Code-Review=+2