Update patch set 5
Patch Set 5:

I meant to ask this earlier, but I couldn't find a test that verifies this functionality always works on bionic. Is there one? And is my assumption that this should always work on bionic true, or does this depend on the kernel being used?

Patch-set: 5
Reviewer: Christopher Ferris <1019050@85c56323-6fa9-3386-8a01-6480fb634889>
Label: Verified=0