Update patch set 3
Patch Set 3:

I'm not particularly attached to this test. I made it to find mistakes which I of course made. Those are fixed already so I will remove the test, copy the method to file where it's used and submit the cl on Monday. Thanks for the reviews guys.

Patch-set: 3