Update patch set 3
Patch Set 3:

sorry, no, i think this is ready now. it certainly fixes gcov and passes the new tests.

i'm still not sure there aren't surprises lurking if the fast path doesn't call __srefill (which it can't), but i plan on coming back to that, trying to write tests that should fail based on inspection of __srefill.

i suspect __srefill probably needs to be refactored into something more like __swsetup (which is called to switch from reading to writing on a stream). i think we want an __srsetup to switch to reading, and a separate __srefill that calls __srsetup but also tries to fill the buffer. i'm not quite sure about the unget buffer, which is one reason i'd rather just come back to this. (the main reason being "the current code is *known* broken and this fixes the known breakage at least".)

Patch-set: 3
Reviewer: Elliott Hughes <1003224@85c56323-6fa9-3386-8a01-6480fb634889>
Label: Verified=0