Update patch set 1
Patch Set 1:

> The only process that's really of interest for DNS lookups is netd,
 > because all app lookups DNS go through it. It's easy to check an
 > environment variable in res_init, but how can we change netd's
 > environment variables without being netd?

i debug on the desktop whenever i can, so i was just assuming you'd set the environment variable on the command-line when running the unit tests.

if you mostly want this for netd... (and i'm assuming you don't regularly restart netd or want it to read a system property each time.) maybe add a command to netd that lets you ask it to turn debugging on/off?

Patch-set: 1