Update patch set 3
Patch Set 3:

There is one drawback to this approach. If the freebsd code calls one of the overriden functions, then the old version is used, not the new version.

I'm not sure this is okay, but I'm putting this up so people can look and see if other than that, this seems okay.

Patch-set: 3