Update patch set 3
Patch Set 3:

> Could you split the python changes and the generated headers into
 > separate patches? Tough to find the parts that actually need to be

Only two .py files (clean_header.py and cpp.py) contain actual changes to be reviewed. All the header files are generated by the python scripts. I'm also including a 'kernel-3.18.diff' file which aggregates all the changes in header files due to the new script. It only gives the idea of all the differences from the old version, not intended to be reviewed (I've examined all the lines in this file).

 > reviewed, plus it will be easier to revert things if needed.
not sure about the "revert" thing here. did you mean to revert the header files to the old versions if something goes wrong? it can be done by reverting the two scripts and run 'update_all.py' again.

Patch-set: 3