Update patch set 1
Patch Set 1:

By the way, I was worried of atomic_compare_exchange_weak_explicit(memory_order_relaxed). If it can't compare
with the newest value or the written value can't be seen by other CPUs, we are in trouble as different 
sem_wait callers may read the same value. But it seems we cpu should ensure LL/SC or compare_exchange
instructions can work in consideration of cache coherency. So we don't need to worry about it as much as
atomic_load(memory_order_relaxed) or atomic_store(memory_order_relaxed).

Patch-set: 1