Update patch set 4
Patch Set 4:

The external discussion is unfortunately in no danger of convergence.  Rich Felker is arguing for sequential consistency, several others against.  The current spec does seem to require seq_cst, in spite of the fact that it doesn't guarantee enough to make that useful.  It remains unclear to me whether performance actually matters that much; most of the usage examples that people generate seem to also involve frequent context switches, in which case an extra fence doesn't really matter.  If we check in the seq_cst version, I think it would be good to add a very visible comment that we reserve the right to change our mind if Posix does.

Patch-set: 4