Update patch set 7
Patch Set 7:

One additional but important note... I see that you are not splitting the exec segment, but still include code to 'relocate' the relocations, adjust symtab, and so on.

The DSO layout places relocation sections before .text.  If you remove relocations and then close up the hole, this changes the start of .text in memory from what it would have been before packing.  Everything still runs -- branches are either pc-relative or fixed up by relocations themselves.  BUT... Dwarf DIE information contains fixed value code addresses that are no longer valid once the hole is closed.  And Breakpad uses DIEs extensively.

We found, then, that this way of working entirely broke breakpad and libbacktrace.  The solution was to split the exec segment into two, so that .text's start does not move:
https://codereview.chromium.org/535943002/

Breakpad needed a small fix to understand that /proc/self/maps may now show more than one exec segment for a mapped library, but other than that it seemed to work okay.  Libbacktrace is still an issue (bug status there is wontfix).

If you don't want to split exec like this to maintain .text start, the only other option I can come up with is to reorder sections in the DSO so that relocations come after .text and other things that DIEs can reference.  This comes with problems of its own.  To do this with ld.bfd requires a linker script, and these are nasty and tricky things that even linker experts and authors have considerable problems getting right, and generally recommend against trying.  And gold does not support linker scripts at all.

Maybe you have a plan for all this.  Worth noting, though, that thanks to things like breakpad, libbacktrace, gdb and so on, there is often very little room for manoeuvre in this area.

Patch-set: 7
Label: Verified=0