Update patch set 1
Patch Set 1:

Especially in the two thread case, I wonder whether we sometimes avoid the futex calls entirely, if the other thread has already done the post by the time we hit the wait?  Is the performance difference really due to exceeding the number of cores, or is it due to occasionally hitting this go-fast mode in which nothing actually suspends?

Patch-set: 1