Update patch set 5
Patch Set 5:

Increasing the ulp fixed all of the failing tests. I'm leaving nextafterf out until the data is made to be float.

There is one caveat to this, the math library says that some of the values can be 2.5 ulps away. Unfortunately, we don't have any way to represent this with the current test system.

As I understand it, 2.5 ulps means that if the exact correct mathematic value is not representable as a float/double, then the computed answer can be up to 2 ulps from the next value larger than the exact answer, and up to 2 ulps from the value smaller than the exact answer. And if the exact answer can be represented, then any value up to 2 ulps away is good.

To represent this for our tests, we'd have to allow two possible expected values for the hyberbolics, which represent the two possible correct values. One of which could be ignored if the exact value can be represented exactly.

Right now, I just set these 2.5 ulps functions to 2 ulps so we are going to be slightly more restrictive than the spec. Please let me know if this is not okay and/or my understanding is incorrect.

Patch-set: 5