Update patch set 10
Patch Set 10: Code-Review+1

When Dalvik was around we had the ability to build and run on a non VFP ARM platform (we had a soft-float variant for all the FLOAT_* and DOUBLE_* opcodes).

This is not the case with ART. We always generate VFP instructions and moreover we use a hard-float ABI (internal to the VM). Thus - do we still need non-VFP handling in bionic and probably other places too?

Patch-set: 10
Reviewer: Serban Constantinescu <1030338@85c56323-6fa9-3386-8a01-6480fb634889>
Label: Code-Review=+1