Update patch set 1
Patch Set 1:

> As far as I can see, this CL is to avoid deadlock between
 > s_tls_map_lock_ and g_thread_list_lock
 > which doesn't happen currently. All uses of s_tls_map_lock_ happen
 > in pthread_key.cpp, so we can
 > avoid the deadlock by always locking g_thread_list_lock before
 > s_tls_map_lock_.

the key is "currently". it's not a lock ordering if you don't enforce it.

 > The situation we need two locks only happens in pthread_key_delete.
 > So locking g_thread_list_lock in
 > ScopedTlsMapAccess may not be necessary and may confuse me. And if
 > we did that, s_tls_map_lock_ can
 > be removed and ScopedTlsMapAccess is really protected by
 > g_thread_list_lock, more confusing?
 > And even we enforce the order of the two locks, we still can't call
 > pthread_key_delete when
 > g_thread_list_lock is held. Because g_thread_list_lock is a normal
 > pthread_mutex_t which doesn't allow
 > recursive lock.

it's our lock so we can make it recursive.

 > If we realy have to lock g_thread_list_lock in
 > ScopedTlsMapAccess, we need to make
 > sure pthread_key_create/pthread_setspecific will not be called when
 > g_thread_list_lock is held.

ideally i wanted to remove the global thread list (lock): https://android-review.googlesource.com/#/c/98889/

but we didn't come up with a way to do that without making pthread_getspecific significantly more expensive. we don't care about the cost of the other pthread_key_t calls because they're rare. but pthread_getspecific needs to be cheap. it should be possible to do that at the cost of extra space, but we never investigated that. i've been hoping that when we have a thread local variable implementation that isn't fixed-size we should be able to do that.

in the short term, just using g_thread_list_lock and removing the tls map lock sounds quite appealing. what's wrong with that?

Patch-set: 1