Update patch set 1
Patch Set 1:

note that we are working towards being able to use the same (bionic) headers in the NDK for all API levels. so when that ships we'll be able to kill a lot of differences like this, and we won't need to support things like _U anywhere any more.

Patch-set: 1