Update patch set 2
Patch Set 2:

It's possible but would split the test cases into different files to compile some with gcc and the rest with clang. So far, we have only these long double cases failing with clang. The other failed cases are stack_unwinding and fchmodat, and they also failed with gcc.

Patch-set: 2