Update patch set 10
Patch Set 10:

By the way, I ran some benchmarks and on an old Nexus 7, flooring of a negative value is ever so slightly slower than the same code. My guess is because Nexus 7 is not extremely sensitive to branches.

It's not for all negative values, but -1.5 does demonstrate this. It's so slight that I don't think it matters though.

I'll do a couple other checks, but I think this is okay. If someone can demonstrate a way this code could generate a floating point exception, please let me know.

Patch-set: 10