Update patch set 4
Patch Set 4:

All bionic tests are now run in a forked process, so I think the CrashCatch is not necessary and makes this a bit more complicated. This means that if one of the tests crashes, it won't stop running the other tests.

Patch-set: 4
Reviewer: Christopher Ferris <1019050@85c56323-6fa9-3386-8a01-6480fb634889>
Label: Verified=0