Update patch set 1
Patch Set 1:

There is one pathological case where this hurts performance that I didn't think was worth fixing. If the internal buffer holds the last bit of a file, and you are reading a large amount of data, and then you fseek backwards within the original internal buffer, you'll have to refill the buffer.

This is such an unlikely thing, that I think this simple change should be okay.

Patch-set: 1