Update patch set 2
Patch Set 2:

Can you explain the motivation here?

Posix calls invocation of pthread_getspecific() with a bad key undefined behavior.  In my experience pthread_getspecific() is commonly used in very performance critical situations.  And we're adding unconditional checking code to the critical path.  I suspect, though don't know, that the liberal Posix spec here was designed to support faster implementations.  We already had much more checking than required.  Do we really need to add more?

The added overhead here is not huge, but I don't think it's completely negligible.  It adds an addition on the critical path, and it prevents the bounds check from being optimized to a single comparison.  (I don't know if the compiler is performing that optimization or not.)

Patch-set: 2