Update patch set 2
Patch Set 2:

> Can you explain the motivation here?

see the linked-to bug.

 > Posix calls invocation of pthread_getspecific() with a bad key
 > undefined behavior.  In my experience pthread_getspecific() is
 > commonly used in very performance critical situations.  And we're
 > adding unconditional checking code to the critical path.  I
 > suspect, though don't know, that the liberal Posix spec here was
 > designed to support faster implementations.  We already had much
 > more checking than required.  Do we really need to add more?
 > 
 > The added overhead here is not huge, but I don't think it's
 > completely negligible.  It adds an addition on the critical path,
 > and it prevents the bounds check from being optimized to a single
 > comparison.  (I don't know if the compiler is performing that
 > optimization or not.)

yabinc: what do the benchmarks say?

i think we can probably make everyone happy though, on the assumption that useful code must call pthread_setspecific (but does so only rarely) --- we can abort if we're given a bad key there, but we could actually remove the validity check completely in pthread_getspecific. that should save us a nanosecond or two, but also make it more likely (than <= L) that incorrect code will crash (if not abort cleanly with an error) because we'll return garbage or SEGV if we just try to use the bad key.


(personally i'd rather set/clear the top bit as a validity marker than add/subtract 1.)

Patch-set: 2