Update patch set 1
Patch Set 1:

Is the lack of noatsecure permission for init to the service domains a problem for any usage of init.environ.rc for global environment setup?  I'm not sure how much we gain by not allowing init noatsecure to the domains to which it transitions given that init is necessarily highly privileged and can reload SELinux policy.

Patch-set: 1