Update patch set 2
Patch Set 2:

The new code passes all of the testing I've done.

However, there are some performance issues. For x86, I measured the performance on a nexus player device with clocks locked to max. For x86_64, I used my desktop system, so it might not apply to a real android system.

For x86 and x86_64:

asin is slower than the current code except for asin(-0.75). I didn't try that many values, is there a range that is expected to be faster?

cos is slower than the current code except for cos(-234.5). As with asin, is there a particular range that is supposed to be faster.

For x86 only:

sin is slower for a lot of values between -1.0 and 1.0.

For these functions (hypot, cos, sin, tan), calling with inf or nan is about eight times slower than the current C code. On most of the other functions, the new code is about twice as slow for these same values. It seems okay to be slower in these special cases.

For x86_64 only:

acos is slower than the current code for all values I tried.
sin is slower than the current code for all values I tried.

tan is slower than the current code for most values, except for tan(-0.92) and tan(0.36) where the new code is faster.

tanh is about the same, except the inf, nan cases are slower than the current code.

As with x86_64, most function special cases for inf, nan are slower, not usually twice as slow, just slightly slower.

Are these issues known?

Patch-set: 2
Label: Verified=0