Update patch set 2
Patch Set 2:

> The details basically look good to me, though I'm not sure I
 > understand the big picture.  Are we removing any locking?

yes, the big picture is "remove the global thread list". it's made some things easier, but it's a single point of contention, and it does lead to lock ordering bugs. (other C libraries don't have a global thread list, presumably for these reasons.) it seems like it's time we did the extra work.

Patch-set: 2