Update patch set 2
Patch Set 2:

my understanding of /oem is that there's not supposed to be any executable code on it, so adding it to the dynamic linker's search path would be a mistake. in which case, we probably want to add a CTS test to ensure that /oem *isn't* on the dynamic linker's default path.

romlem?

Patch-set: 2