Update patch set 2
Patch Set 2:

Elliott: Should we also fix up getauxval in bionic/libc/private/KernelArgumentBlock.h so it has the same behavior?

Also, we might want to fixup the AT_SECURE handling in the dynamic linker to check to make sure AT_SECURE really is set, and not just missing?

Patch-set: 2