Update patch set 1
Patch Set 1:

It is surprising that we didn't notice this. After checking a little, I found that the strings like "u1_root", "nobody" are stored together. On N9, the string seems to be padded to 16-byte alignment, while the string is not padded on x86. This makes a difference in test result.

Patch-set: 1