Update patch set 2
Patch Set 2:

The memchr function is used in a fair bit of third party code so once there's support for using this in the NDK it will be useful. The reverse variant is rarely used but there's not really a maintenance burden as it's pretty much the same code.

These aren't nearly as important as stuff like memcpy, fread, fwrite, etc. but I still think it's worth covering.

Patch-set: 2