Update patch set 3
Patch Set 3:

> I don't think it's essential. I'd rather have a clean
 > implementation than catch something at compile time.

especially because __realpath_chk_fail is a turd that's only there because of a clang limitation. this stuff is ABI that we're going to have to live with forever. __realpath_chk seems to leave more room for maneuver if we are going to add something.

alternatively, let's plan ahead and avoid having 50 more of these special purpose "abort with a message" functions. it's far more cost-effective for us to just have one __fortify_chk_fail or whatever in that case.

Patch-set: 3
Reviewer: Elliott Hughes <1003224@85c56323-6fa9-3386-8a01-6480fb634889>
Label: Verified=0