Update patch set 3
Patch Set 3:

Since everything here *can* be caught at compile-time, any addition to the ABI for this will become obsolete if and when Clang implements the error attribute.

It might be best to just leave out Clang support for this fortified function for now so it can just have a simple 4-line implementation in the header with no ABI addition.

Patch-set: 3