Update patch set 2
Patch Set 2:

i think if you want to go this route, you need to make it impossible to call snprintf from the linker, because it'll creep back in again.

the thing that really worries me, though, is that there's nothing special about snprintf --- it's any function that can transitively access a thread-local variable [that's implemented with pthread_key_create rather than a well-known TLS slot]. that makes me worry that we need a real solution. possibly even to the extent of disallowing pthread_key_create in libc.

Patch-set: 2
Label: Verified=0