Update patch set 1
Patch Set 1:

We should check this on individual CPUs - i.e. A53 only, A57 and so on. I can give it a go on Denver (Nexus 9). If this patch still bring a performance improvement we can probably add a arch-arm64/cortex-a53/ and a arch-arm64/cortex-a57. Though if the benefit is not substantial on these microarchitectures it might not be worth adding all this.

Patch-set: 1
Reviewer: Serban Constantinescu <1030338@85c56323-6fa9-3386-8a01-6480fb634889>
Label: Code-Review=0