Update patch set 2
Patch Set 2:

Adding to Nikola's comments... DT_DEBUG can be used if the dynamic section is writable. In the past we tried using DE_DEBUG for MIPS on Android by maintaining a local patch to gcc and allowing the dynamic section to be writable. Otherwise on MIPS, the dynamic section was always read only. Recently, I read in the below mail thread as to why MIPS dynamic section was chosen to be read only. 
https://sourceware.org/ml/binutils/2014-10/msg00203.html

Patch-set: 2
Reviewer: Raghu Gandham <1064548@85c56323-6fa9-3386-8a01-6480fb634889>
Label: Verified=0