Update patch set 1
Abandoned

After further testing, this change isn't really worth it. There are more cases where this has a speed penalty than an improvement, and that penalty is around 15%, so if you did a big loop through a bunch of various length strings with and without this change it would probably take the same amount of time to complete, but ldmia would have a consistent speed across the board, which is far more preferrable with something like strings; if this was something like memcpy, where I think values are usually more aligned, then it would maybe be worth it.

Patch-set: 1
Status: abandoned