Update patch set 1
Patch Set 1:

I verified that for a system with dlmalloc, the symbols are exported and the dlmalloc function call into the real calls.

I also verified for a system with jemalloc, the symbols are exported.

I did this for 32 bit and 64 bit for both sets (using flounder for with dlmalloc, and arm64 for with jemalloc).

Patch-set: 1