Update patch set 1
Patch Set 1:

Dimitry, could you please forward the contents of http://b/22045439, and the name of the mips32 programs you tried, and how that mips32 program was compiled?

I am guessing that the mips32 program failed with linker message "Kernel or cpu failed to set FRE mode required for running ...".

That particular issue is expected, and will resolve when we update aosp's copy of the mips kernel (alternatively, if the program is recompiled with current mips compilers).

If this test failure was some other issue, I'd like your example to work on it further.

thanks,
    Duane

Patch-set: 1