Update patch set 1
Patch Set 1:

This patch is not yet fully complete, but I think it can be merged, assuming it passes the tests :-).

If you guys want me to approach the problem in  a different way, please let me know. I am more then happy to re-dress the code any which way preferred. I believe the way I am approaching this is likely to have the least impact to the code, but whatever ... when in Rome ... :-)

Patch-set: 1
Reviewer: Jason Kim <1076843@85c56323-6fa9-3386-8a01-6480fb634889>
Label: Verified=0