Update patch set 7
Patch Set 7:

Ah, yeah. It isn't the first time that's been an issue.

I think it might be a bug with Clang's implementation of GNU inline semantics in C++, as GCC determines the defacto standard. Clang implements it the same way as GCC for C so it wasn't needed for cases like fwrite/fread (which is good, because moving those out of upstream-openbsd would be annoying).

Patch-set: 7