Update patch set 2
Patch Set 2:

I switched the problematic static check to a warning on 64-bit to work around the problem. FWIW, I was able to replicate a similar issue on x86_64 and it doesn't happen with GCC's master branch (but I have no idea which commit(s) fixed it and don't really want to bisect).

Patch-set: 2