Update patch set 1
Patch Set 1:

No, the compiler can generate functions that do this branch. In fact, the code in question is in C++, there is no assembler.

There are probably other cases where the line number would be wrong if we don't adjust the ip. Amongst them, yes if any branch to another function occurs that isn't expected to return and that is the last instruction in the function.

Patch-set: 1