Update patch set 2
Patch Set 2:

Nice work Jake!

Any luck on getting the Android uprobes profiles? It might turn out that Android prefers 8b memcpy() and in that case the original is probably the one to go for. I have not looked yet at the code in detail, but judging by the throughput it seems that for 1k, 32k and 64k on MSM8974 hits a CPU limitation (in the original code). Branch prediction went wrong :-) ? Maybe we can fix the original to start of with an then work from there. That should be easier to review... at least.

Patch-set: 2
Reviewer: Serban Constantinescu <1072549@85c56323-6fa9-3386-8a01-6480fb634889>
Label: Verified=0