Update patch set 5
Patch Set 5:

I fixed the initialization of main thread stack so that safestack can be used in .preinit_array and also for static executables.

I've added a test.

Please consider this CL for commit. All new functionality is under ro.libc.safestack, which, as of now, is never set.

Patch-set: 5
Label: Verified=0