Update patch set 10
Patch Set 10:

(8 comments)

if we want to make this non-experimental, it'll have to be based on a BoardConfig.mk variable so we can turn it on on a per-device basis. (in the absence of very detailed and convincing evidence that there's no memory cost, no performance cost, and no compatibility cost.) we're also going to have to work seriously to try to get numbers for those three things.

do you have a rough idea of when the necessary clang changes will be in google3's clang? (as a lower bound for when they'll make it to Android's clang.)

independently, we'll have to run this by FC. i'll talk to bdc about that.

Patch-set: 10