Update patch set 3
Patch Set 3:

I've verified that this keeps the source lists the same on all default lunch targets (internal and aosp), and a number of other variants I've tested. With one exception -- x86 was putting some upstream-(free|open)bsd/... files in libc_common instead of libc_freebsd/openbsd, and it put the __chk files in libc_common_src_files instead of libc_bionic. This makes it the same as the other architectures

Patch-set: 3
Label: Code-Review=0