Update patch set 9
Patch Set 9:

> > Hey guys, seriously, whats the hold up? :-)
 > 
 > Jason, the team asked you directly to fix the malloc->mmap issue
 > first. This patch is still too large for anyone to review. It would
 > also be great to understand whether/how OpenBSD works with TSan. If
 > it works properly, we should probably understand how it works.


Hi guys,

This change does contain the first part for mmap() (not malloc(), as its a bit more involved).

The files I touch in this patch is because they all "belong together".

If we really wanted, please just take a look at 

	libc/private/bionic_asm.h

That can be merged separately (FIRST)
and each of the .S files (with mostly 1 line changes) can be merged afterwards. Though if there is a compile error somwehre, it might be best to put this patch in as is.


This patch  really is about as small as I can make it and still have it be a single coherent unit of thought.

Its not big! Really! (please just spend 30 seconds and click through the patch ;-)

Patch-set: 9