Update patch set 1
Patch Set 1:

> The linker's own executable code will be no more random, and I'd
 > expect that there are more than enough ROP gadgets for an attacker
 > in there.

Yes this does not improve randomness of placement of linker's own executable code as well as the code of the the main executable. We are not really in control of these processes here - kernel have already done the work for us.

 > Each dlopen'ed library is also going to introduce a
 > separate, even lower entropy (usually 0 bits) area of executable
 > code and even libc itself dynamically loads 2 libraries after the
 > initial linking. 

Not sure I follow this, why?

Patch-set: 1
Label: Verified=0