Update patch set 2
Patch Set 2:

It's probably worth doing, but I don't think it's as much of an improvement as it seems. It will make it a little bit less reliable to attack say, global function pointers in an initially loaded library. I don't think it makes ROP harder in practice though.

Patch-set: 2