Update patch set 2
Patch Set 2:

> I did some testing of this patch, and I'm seeing *more* randomness
 > than what I would have expected. That's good, although I don't
 > quite understand it...

The linker randomized pretty much the whole graph of not loaded dt_needed libraries (not only the first level). The bigger the graph - the better randomization.. mediaserver has a lot of dt_needed libraries so this is why we get good results.

Note that the placement of libc.so depends not only on the order number but also on the sizes of previously loaded libraries.

Patch-set: 2