Update patch set 2
Patch Set 2:

> Abandoned
 > 
 > Sorry, I didn't realize that it is not in posix standard.

i could be persuaded that we should add it anyway, though. annoyingly i notice that glibc and Mac OS do things differently from one another; Mac OS seems to have pthread_yield_np instead.

there are a handful of references in the tree.

but if we do add it, it should probably just be an alias.

Patch-set: 2