Update patch set 6
Patch Set 6: Code-Review+1

William: The purpose of this change is to add read access control to the property space. We already have write access control, but lacking read access control is a gap we're trying to fix.

Moving the property space out of bionic is a good idea, but a significant undertaking and should be out of scope for this change.

Patch-set: 6
Label: Code-Review=+1