Update patch set 8
Patch Set 8:

(1 comment)

The N9 issue is why I have the code to add the debug_prop properties to /dev/__properties__.  It looks like something on the device is expecting them to be there.  Adding those properties works around the issue, but isn't something we'd want to do.  I haven't been able to determine the culprit.

Patch-set: 8