Update patch set 11
Patch Set 11:

(1 comment)

The implementation looks good to me.  I'm not positive the number of fences is quite optimal, but that's probably in the noise without adding a spin loop.  We'll usually have to enter the kernel anyway.

I would still like to see the added test I suggested last time.  This is tricky enough that I think we need a more realistic test before inflicting this on the world.

Patch-set: 11
Label: Verified=0