Update patch set 9
Patch Set 9:

(1 comment)

Nick - I talked to Elliott today and he said that we should not need to support the case of <older application/library statically linked to libc> on <Android versions after this change>.  This means that we can re-purpose /dev/__properties__ completely.  

My idea here is to use /dev/__properties__ as the directory that contains the secure property files in the new design.  I had previous changes that did this and it should limit the amount of changes to be made.

That said, we'll need a place that any process that accesses properties can access to store the serial numbers.  Can we make sure a guarantee with the 'default_prop' context or some other existing context?  If not, can we create such a context for this purpose?

I'll push code to do this later today for review.

Patch-set: 9
Label: Verified=0