Update patch set 23
Patch Set 23: Code-Review+2

sgtm. security labels more than even just 40 characters are going to be enough of a pain for the humans reading them that i find it hard to believe we're actually going to find 1024 limiting. (and it's an internal implementation limit anyway, so we can easily change it if there's something we haven't thought about.)

Patch-set: 23
Label: Code-Review=+2