Update patch set 1
Patch Set 1:

Is this ready for review? Would this fail for brillo since ndk_cruft is the only thing that defines __sinit, and some of the openbsd code still references __sinit?

I know that because of the define of __sdidinit, the __sinit will never execute, but I thought one of compilers still required that you had the function definition, but I might be remembering incorrectly.

Patch-set: 1