Update patch set 1
Patch Set 1:

> Interesting. This might be what I was hitting before. I never
 > narrowed down the exact cause after I found a workaround (didn't
 > have an exact idea of where it crashed, only that it was in the
 > init code) and misidentified the cause of the issue.

the dmesg line that tells you init died gives you an eip value. you can objdump --disassemble the init binary in symbols/ and then search for that address. it would be helpful in more complex cases if the kernel would unwind, but in this case there's only one caller of __set_tls anyway.

Patch-set: 1